Add wofi to sway setup
The commonly used dmenu is annoying to deal with because it works via
xwayland. This causes issues when the window loses focus and it has
weird interactions with waybar. wofi works much better. I'm not sure if
I'll stick to it or find another alternative but let's make the change
for now.