Update gnome setup
Use ubuntu's appindicator instead of topicon. The latter is
unfortunately unmaintained (or loosely maintained?). It also doesn't
look very "native". It probably lets the app draw its own window and I
suspect it's even using xwayland to do so. Among other issues, this
sometimes makes it slow and stuttery. Unfortunately appindicator is not
as customizable so it's not a strict improvement, but it's pretty good.

Use dash-to-panel to get a more traditional desktop flow, similar to how
ubuntu does it.

Replace gnome-terminal with kitty.