Use official zsh-theme-powerlevel10k
Replace the git version with this. We used to use the git version
because the official package had a record of breaking, but there seems
to have been an effort to improve it in recent years and it's time to
try it out again and attempt to get rid of an AUR dependency. Note that
since the git version provides zsh-theme-powerlevel10k anyway, we can
locally switch to the git version without modifying the PKGBUILD if we
want to.

One thing to note is that due to a disagreement with how packages expect
to use CMAKE_BUILD_TYPE and how upstream expects it to be used, it
appears that the official package is built without optimizations. If
this appears to be an issue, we can always build locally by adding
`-DCMAKE_BUILD_TYPE=None` to the official PKGBUILD.